+++ app/controllers/site_users_controller.rb
@@ -1,3 +1,4 @@
+# Allows members to invite others to collaborate on a project as a site member.
 class SiteUsersController < ApplicationController
   before_action :authenticate_user!, except: [:invite]
 
@@ -5,7 +6,8 @@ class SiteUsersController < ApplicationController
     session[:site_invite_token] = params[:site_invite_token]
     if current_user
       site_invite_token = session[:site_invite_token]
-      if @site_user = SiteUser.find_by_invite_token(site_invite_token)
+      @site_user = SiteUser.find_by_invite_token(site_invite_token)
+      if @site_user
         redirect_to accept_project_site_users_path(@site_user.project)
       else
         session[:site_invite_token] = nil
@@ -21,7 +23,7 @@ class SiteUsersController < ApplicationController
     @site_user = SiteUser.current.find_by_id(params[:id])
     @site = current_user.all_sites.find_by_id(@site_user.site_id) if @site_user
 
-    if @site and @site_user
+    if @site && @site_user
       @site_user.send_invitation
       # resend.js.erb
     else
@@ -32,40 +34,36 @@ class SiteUsersController < ApplicationController
   def accept
     site_invite_token = session.delete(:site_invite_token)
     @site_user = SiteUser.find_by_invite_token(site_invite_token)
-    if @site_user and @site_user.user == current_user
+    if @site_user && @site_user.user == current_user
       redirect_to [@site_user.site.project, @site_user.site], notice: "You have already been added to #{@site_user.site.name}."
-    elsif @site_user and @site_user.user
-      redirect_to root_path, alert: "This invite has already been claimed."
+    elsif @site_user && @site_user.user
+      redirect_to root_path, alert: 'This invite has already been claimed.'
     elsif @site_user
       @site_user.update_attributes user_id: current_user.id
-      redirect_to [@site_user.project, @site_user.site], notice: "You have been successfully been added to the site."
+      redirect_to [@site_user.project, @site_user.site], notice: 'You have been successfully been added to the site.'
     else
       redirect_to root_path, alert: 'Invalid invitation token.'
     end
   end
 
   # DELETE /site_users/1
-  # DELETE /site_users/1.json
   def destroy
     @site_user = SiteUser.current.find_by_id(params[:id])
     @site = current_user.all_sites.find_by_id(@site_user.site_id) if @site_user
     @project = @site_user.project if @site_user
 
     respond_to do |format|
-      if @site and @project
+      if @site && @project
         @site_user.destroy
         format.html { redirect_to [@site.project, @site] }
-        format.json { head :no_content }
         format.js { render 'projects/members' }
-      elsif @site_user.user == current_user and @project
+      elsif @site_user.user == current_user && @project
         @site = @site_user.site
         @site_user.destroy
         format.html { redirect_to root_path }
-        format.json { head :no_content }
         format.js { render 'projects/members' }
       else
         format.html { redirect_to root_path }
-        format.json { head :no_content }
         format.js { render nothing: true }
       end
     end