+++ db/migrate/20151005142207_add_blinding_enabled.rb
@@ -1,0 +1,5 @@
+class AddBlindingEnabled < ActiveRecord::Migration
+  def change
+    add_column :projects, :blinding_enabled, :boolean, null: false, default: false
+  end
+end