+++ db/migrate/20151005142952_add_unblinded_to_project_users_and_site_users.rb
@@ -1,0 +1,6 @@
+class AddUnblindedToProjectUsersAndSiteUsers < ActiveRecord::Migration
+  def change
+    add_column :project_users, :unblinded, :boolean, null: false, default: true
+    add_column :site_users, :unblinded, :boolean, null: false, default: true
+  end
+end