+++ db/migrate/20151005144939_add_only_unblinded_to_designs.rb
@@ -1,0 +1,5 @@
+class AddOnlyUnblindedToDesigns < ActiveRecord::Migration
+  def change
+    add_column :designs, :only_unblinded, :boolean, null: false, default: false
+  end
+end