+++ CHANGELOG.md
@@ -18,6 +18,10 @@
   - Sheets can be filled out and added to an adverse event
   - Users receive a notification if they have viewable adverse events that have new updates, comments, files, or forms
   - Unblinded project editors and project owner are notified when an adverse event is reported
+- **Blinding Module Added**
+  - Projects can now enable blinding and unblinding of project and site members
+  - Designs are set to be viewable by all, or only to be viewable by unblinded members
+  - Creation of randomizations and adverse event reports is limited to unblinded members
 - **Gem Changes**
   - Updated to pg 0.18.3
   - Removed minitest-reporters